+++ dinosaurs.json
@@ -1,1 +1,1 @@
-[{"name":"Ducky","type":"saurolophus"},{"name":"Spike","type":"stegosaurus"},{"name":"Barney","type":"Purple Dino"}]
+[{"name":"Ducky","type":"saurolophus"},{"name":"Spike","type":"Stegosaurus"},{"name":"Barney","type":"Purple Dino"}]
