+++ views/dinosaurs/prehistoric.ejs
@@ -1,0 +1,5 @@
+<ul>
+    <img src="<%= document.data.illustration.url %>" />
+<span class="image-caption"><%= document.data.type %></span>
+        <% }) %>ç
+</ul>
