+++ views/layout.ejs
@@ -19,6 +19,9 @@
             <li>
                 <a href="/dinosaurs/new"> new dino</a>
             </li>
+            <li>
+                <a href="/dinosaurs/prehistoric"> prehistoric creatures</a>
+            </li>
         </ul>
     </nav>
     <h1> I LOVE DINOSSS!!!! 🦕</h1>
